requirements.txt
# Face Authentication Attendance System
# Python Dependencies

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Database
supabase==2.3.0

# Face Recognition
face-recognition==1.3.0
# dlib==19.24.2  # Install from wheel: pip install dlib-19.22.99-cp310-cp310-win_amd64.whl

# Image Processing
opencv-python==4.9.0.80
numpy==1.26.3
scipy==1.12.0
Pillow==10.2.0

# Environment Variables
python-dotenv==1.0.0

# Development Tools
gunicorn==21.2.0

# Note: dlib requires cmake to build
# On Windows, you may need to install Visual Studio Build Tools
# On macOS: brew install cmake
# On Ubuntu: sudo apt-get install cmake

# Installation order matters for face-recognition:
# 1. pip install cmake
# 2. pip install dlib
# 3. pip install face-recognition
